+++ wouldyourather/src/index.css
@@ -1,4 +1,5 @@
 *{
     margin: 0;
     padding: 0;
+    font-family: 'Open Sans';
 }
